+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [0.5.3]
 - Fix Item's PrefabInfoGroup not loading correctly at Simulator mode.
+- Change some builtin items' size to match real world size.
+- MaterialController will not clone material unless Config has changed.
 
 ## [0.5.2]
 - Update Top Menu.